# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GeneFunk Character Sheet API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: characters
    description: Character sheet operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /characters:
    get:
      operationId: getCharacters
      tags: [characters]
      summary: List all characters for the current user
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharacterSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterBody"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
      required:
        - id

    GetCurrentAuthUserResponse:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
        isAuthenticated:
          type: boolean
      required:
        - isAuthenticated

    CharacterSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        race:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        level:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - level
        - updatedAt

    Character:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        race:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        subclass:
          type: ["string", "null"]
        background:
          type: ["string", "null"]
        level:
          type: integer
        experiencePoints:
          type: integer
        alignment:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        eyes:
          type: ["string", "null"]
        skin:
          type: ["string", "null"]
        hair:
          type: ["string", "null"]
        backstory:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        strength:
          type: integer
        dexterity:
          type: integer
        constitution:
          type: integer
        intelligence:
          type: integer
        wisdom:
          type: integer
        charisma:
          type: integer
        maxHitPoints:
          type: integer
        currentHitPoints:
          type: integer
        temporaryHitPoints:
          type: integer
        armorClass:
          type: integer
        speed:
          type: integer
        initiative:
          type: integer
        hitDice:
          type: ["string", "null"]
        deathSaveSuccesses:
          type: integer
        deathSaveFailures:
          type: integer
        inspiration:
          type: boolean
        proficiencyBonusOverride:
          type: ["integer", "null"]
        savingThrowProficiencies:
          type: array
          items:
            type: string
        skillProficiencies:
          type: array
          items:
            type: string
        skillExpertise:
          type: array
          items:
            type: string
        languages:
          type: ["string", "null"]
        toolProficiencies:
          type: ["string", "null"]
        armorProficiencies:
          type: ["string", "null"]
        weaponProficiencies:
          type: ["string", "null"]
        attacks:
          type: array
          items:
            $ref: "#/components/schemas/AttackEntry"
        spells:
          type: array
          items:
            $ref: "#/components/schemas/SpellEntry"
        spellcastingAbility:
          type: ["string", "null"]
        spellSaveDC:
          type: ["integer", "null"]
        spellAttackBonus:
          type: ["integer", "null"]
        spellSlots:
          type: object
          additionalProperties:
            type: object
            properties:
              total:
                type: integer
              used:
                type: integer
            required:
              - total
              - used
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentEntry"
        currency:
          type: object
          properties:
            cp:
              type: integer
            sp:
              type: integer
            ep:
              type: integer
            gp:
              type: integer
            pp:
              type: integer
          required:
            - cp
            - sp
            - ep
            - gp
            - pp
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureEntry"
        geneMods:
          type: array
          items:
            $ref: "#/components/schemas/GeneModEntry"
        cybernetics:
          type: array
          items:
            $ref: "#/components/schemas/CyberneticEntry"
        conditions:
          type: array
          items:
            type: string
        exhaustionLevel:
          type: integer
        passivePerception:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - level
        - strength
        - dexterity
        - constitution
        - intelligence
        - wisdom
        - charisma
        - maxHitPoints
        - currentHitPoints
        - temporaryHitPoints
        - armorClass
        - speed
        - initiative
        - deathSaveSuccesses
        - deathSaveFailures
        - inspiration
        - experiencePoints
        - exhaustionLevel
        - savingThrowProficiencies
        - skillProficiencies
        - skillExpertise
        - attacks
        - spells
        - spellSlots
        - equipment
        - currency
        - features
        - geneMods
        - cybernetics
        - conditions
        - createdAt
        - updatedAt

    AttackEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        attackBonus:
          type: ["string", "null"]
        damage:
          type: ["string", "null"]
        damageType:
          type: ["string", "null"]
        range:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - name

    SpellEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        level:
          type: integer
        school:
          type: ["string", "null"]
        castingTime:
          type: ["string", "null"]
        range:
          type: ["string", "null"]
        components:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        prepared:
          type: boolean
      required:
        - id
        - name
        - level
        - prepared

    EquipmentEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
        weight:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        equipped:
          type: boolean
      required:
        - id
        - name
        - quantity
        - equipped

    FeatureEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        source:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - id
        - name

    GeneModEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - id
        - name
        - active

    CyberneticEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slot:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - id
        - name
        - active

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: ["string", "null"]
      required:
        - code
        - code_verifier
        - redirect_uri
        - state

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    CreateCharacterBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    UpdateCharacterBody:
      type: object
      properties:
        name:
          type: string
        race:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        subclass:
          type: ["string", "null"]
        background:
          type: ["string", "null"]
        level:
          type: integer
        experiencePoints:
          type: integer
        alignment:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        eyes:
          type: ["string", "null"]
        skin:
          type: ["string", "null"]
        hair:
          type: ["string", "null"]
        backstory:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        strength:
          type: integer
        dexterity:
          type: integer
        constitution:
          type: integer
        intelligence:
          type: integer
        wisdom:
          type: integer
        charisma:
          type: integer
        maxHitPoints:
          type: integer
        currentHitPoints:
          type: integer
        temporaryHitPoints:
          type: integer
        armorClass:
          type: integer
        speed:
          type: integer
        initiative:
          type: integer
        hitDice:
          type: ["string", "null"]
        deathSaveSuccesses:
          type: integer
        deathSaveFailures:
          type: integer
        inspiration:
          type: boolean
        proficiencyBonusOverride:
          type: ["integer", "null"]
        savingThrowProficiencies:
          type: array
          items:
            type: string
        skillProficiencies:
          type: array
          items:
            type: string
        skillExpertise:
          type: array
          items:
            type: string
        languages:
          type: ["string", "null"]
        toolProficiencies:
          type: ["string", "null"]
        armorProficiencies:
          type: ["string", "null"]
        weaponProficiencies:
          type: ["string", "null"]
        attacks:
          type: array
          items:
            $ref: "#/components/schemas/AttackEntry"
        spells:
          type: array
          items:
            $ref: "#/components/schemas/SpellEntry"
        spellcastingAbility:
          type: ["string", "null"]
        spellSaveDC:
          type: ["integer", "null"]
        spellAttackBonus:
          type: ["integer", "null"]
        spellSlots:
          type: object
          additionalProperties:
            type: object
            properties:
              total:
                type: integer
              used:
                type: integer
            required:
              - total
              - used
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentEntry"
        currency:
          type: object
          properties:
            cp:
              type: integer
            sp:
              type: integer
            ep:
              type: integer
            gp:
              type: integer
            pp:
              type: integer
          required:
            - cp
            - sp
            - ep
            - gp
            - pp
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureEntry"
        geneMods:
          type: array
          items:
            $ref: "#/components/schemas/GeneModEntry"
        cybernetics:
          type: array
          items:
            $ref: "#/components/schemas/CyberneticEntry"
        conditions:
          type: array
          items:
            type: string
        exhaustionLevel:
          type: integer
        passivePerception:
          type: ["integer", "null"]