# lib/api-zod/src/generated/types/attackEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeneFunk Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface AttackEntry {
  id: string;
  name: string;
  /** @nullable */
  attackBonus?: string | null;
  /** @nullable */
  damage?: string | null;
  /** @nullable */
  damageType?: string | null;
  /** @nullable */
  range?: string | null;
  /** @nullable */
  notes?: string | null;
}