# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeneFunk Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
import type { AttackEntry } from "./attackEntry";
import type { CharacterCurrency } from "./characterCurrency";
import type { CharacterSpellSlots } from "./characterSpellSlots";
import type { CyberneticEntry } from "./cyberneticEntry";
import type { EquipmentEntry } from "./equipmentEntry";
import type { FeatureEntry } from "./featureEntry";
import type { GeneModEntry } from "./geneModEntry";
import type { SpellEntry } from "./spellEntry";

export interface Character {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  race?: string | null;
  /** @nullable */
  class?: string | null;
  /** @nullable */
  subclass?: string | null;
  /** @nullable */
  background?: string | null;
  level: number;
  experiencePoints: number;
  /** @nullable */
  alignment?: string | null;
  /** @nullable */
  playerName?: string | null;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  eyes?: string | null;
  /** @nullable */
  skin?: string | null;
  /** @nullable */
  hair?: string | null;
  /** @nullable */
  backstory?: string | null;
  /** @nullable */
  notes?: string | null;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  maxHitPoints: number;
  currentHitPoints: number;
  temporaryHitPoints: number;
  armorClass: number;
  speed: number;
  initiative: number;
  /** @nullable */
  hitDice?: string | null;
  deathSaveSuccesses: number;
  deathSaveFailures: number;
  inspiration: boolean;
  /** @nullable */
  proficiencyBonusOverride?: number | null;
  savingThrowProficiencies: string[];
  skillProficiencies: string[];
  skillExpertise: string[];
  /** @nullable */
  languages?: string | null;
  /** @nullable */
  toolProficiencies?: string | null;
  /** @nullable */
  armorProficiencies?: string | null;
  /** @nullable */
  weaponProficiencies?: string | null;
  attacks: AttackEntry[];
  spells: SpellEntry[];
  /** @nullable */
  spellcastingAbility?: string | null;
  /** @nullable */
  spellSaveDC?: number | null;
  /** @nullable */
  spellAttackBonus?: number | null;
  spellSlots: CharacterSpellSlots;
  equipment: EquipmentEntry[];
  currency: CharacterCurrency;
  features: FeatureEntry[];
  geneMods: GeneModEntry[];
  cybernetics: CyberneticEntry[];
  conditions: string[];
  exhaustionLevel: number;
  /** @nullable */
  passivePerception?: number | null;
  createdAt: Date;
  updatedAt: Date;
}