# lib/api-zod/src/generated/types/equipmentEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeneFunk Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentEntry {
  id: string;
  name: string;
  quantity: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  description?: string | null;
  equipped: boolean;
}