# lib/api-zod/src/generated/types/geneModEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GeneFunk Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneModEntry {
  id: string;
  name: string;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  description?: string | null;
  active: boolean;
}